+++ src/types/ShellCheckArchive.ts
@@ -1,0 +1,4 @@
+/**
+ * ShellCheck archive.
+ */
+export type ShellCheckArchive = 'tar.xz' | 'zip';