+++ vitest.config.ts
@@ -1,0 +1,17 @@
+import tsconfigPaths from 'vite-tsconfig-paths';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    setupFiles: ['./tests/vitest.setup.ts'],
+    typecheck: {
+      enabled: true,
+      tsconfig: './tests/tsconfig.json'
+    },
+    coverage: {
+      include: ['**/src/**/*.?(c|m)[jt]s?(x)']
+    }
+  },
+  plugins: [tsconfigPaths()]
+});